+++ src/main/scala/sangria/validation/rules/OverlappingFieldsCanBeMerged.scala
@@ -264,8 +264,8 @@ class OverlappingFieldsCanBeMerged extends ValidationRule {
         typeRes.flatten match {
           case s @ Some(_) ⇒ s
           case None ⇒
-            val type1 = def1 map (d ⇒ ctx.typeInfo.getNamedType(d.fieldType))
-            val type2 = def2 map (d ⇒ ctx.typeInfo.getNamedType(d.fieldType))
+            val type1 = def1 map (d ⇒ d.fieldType.namedType)
+            val type2 = def2 map (d ⇒ d.fieldType.namedType)
             val conflicts = findConflictsBetweenSubSelectionSets(areMutuallyExclusive, type1.asInstanceOf[Option[CompositeType[_]]], ast1, type2.asInstanceOf[Option[CompositeType[_]]], ast2)
 
             subfieldConflicts(conflicts, outputName, ast1, ast2)
@@ -318,7 +318,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule {
         fragmentName1: String,
         fragmentName2: String,
         areMutuallyExclusive: Boolean): Unit = {
-      (ctx.fragments.get(fragmentName1), ctx.fragments.get(fragmentName2)) match {
+      (ctx.doc.fragments.get(fragmentName1), ctx.doc.fragments.get(fragmentName2)) match {
         case (None, _) | (_, None) ⇒ // do nothing
 
         case (Some(f1), Some(f2)) if f1.name == f2.name ⇒
@@ -353,7 +353,7 @@ class OverlappingFieldsCanBeMerged extends ValidationRule {
         fieldMap: MutableMap[String, ListBuffer[AstAndDef]],
         fragmentName: String,
         areMutuallyExclusive: Boolean): Unit = {
-      ctx.fragments.get(fragmentName) match {
+      ctx.doc.fragments.get(fragmentName) match {
         case Some(fragment) ⇒
           val (fieldMap2, fragmentNames2) = getReferencedFieldsAndFragmentNames(fragment)